feat: env's half implementation
working on an env struct to control and organize scopes. had
some issues with getting mutable references working and i realized
it wasn't the safest thing to do (according to the compiler) so i;ll
have to rework it but with a que instead, executing each line and then
seeing what the result is. if its an assignment it will have to add it
to the que and push it upward until it gets to a point where the
assignment can be made (local variables vs global variables and what
scope they are in).

this doesn't build or work currently just an inplace commit for keep
tract of the work and what i was working on + some notes on the intent.