feat: cleans up testing code, adds local macros
adds local macros for easier creation of internal objects
to test if the matching and creation is working correctly.

also make the creation functions for each of the grammar types
consistent, so each has a `create` which makes itself and a
`create_into_gram` which wraps it in a gram.