feat: adds not, local to unary parsing
needed to deal with local variables some how, and decided to
make local an unary function. In `local bob = 4` `local` is modifiying
bob and letting the compiler / interpreter know that bob is a new
variable created inside this scope, and not any other variable named
bob that might be in any of the parent scopes.

also updated the `Unary::create` function because it was hard coded to
only accept one operator, `-`, so now it looks at the OPERATIONS_ORDER
constant (where the operator order is defined) and checks if the
operator is in that 2D array, if it is then it is a proper unary.