test: adds test for chaining binops
the test originally failed because it was matching on
the unary `-` even though it shouldn't have because the
token before the `-` was an expression.

    5 + 4 - 3

The above was matching `-3` as a unary where it actually
should have been `4 - 3` as a binary. to fix this changed
`Parser::peek_expr` to `Parser::peek_expr_before` and
`Parser::peek_expr_after` so it will be able to check if
there is an expression before the current token.

i chose to use a `_before` and `_after` instead of just
passing `pos+1` and `pos-1` because i didn't want to do
a check in the parsing function to make sure that pos != 0
and `pos-1`, `0-1` would fault the process.