feat: implements oop for binary
added a oop function fo `TokenType` that can be used to check how an
operator relates to another operator for their order of operations.
currently this is only setup for binary, because that is the most
complex relation. all unary operators are the same so there is no need
to have one for those.