chore: organizes into sections / subprojects
moved all the main library stuffs to `lua-interpreter` crate and moved the
testing support to the `lua-test` crate. The intent is `lua-test` will
house lua code with equivalent macro! buildings for testing, removing it
from the library code and putting it all in one place.

currently nothing is done, just the organization.