feat: started on the parser
started the inital work on the parser, struggling to find a
way to get this to work easily and cleanly. started with the
binary operations and i think i am going down a direction that
might work. need to refactor what i have done to make it cleaner
because it will get crowded soon.