feat: creates 1 complete path
implements block and chunk, and does the
complete parsing lifecycle for the simple
statement `bob = 4 + -3` or anything
equivalent.

found out i have an issue, but will address
that later, when i redo the error reporting
i need to change up the fundamental way it
works because we destroy the code data for
elements as we combine them, so if we get
to a point (and we will) in the future where
we need to call out a variable somewhere, we
don't have that data anymore because it has been
destroyed during the parsing and scanning phase
such that we can't have access to it anymore
during the eval phase.