feat: adds prefixexp
tweaked `varlist` so that it will only match when looking at the left of
an equals sign. this was giving me problems because i'm using the order
of the lua spec to do the order of parsing. the order there is `varlist`
before `prefixexp` which means that you can never have a `var` on the
right side of an equals sign because a `var` will always be parsed as a
`varlist` before it could be parsed as an `prefixexp`.

added some usize error checking on the final_compress return
(`parser.rs`) because i was getting back code references that were `0,0`
which caused a subtract overflow when trying to get the error message
code references.

adds implementation of `prefixexp` for 2 of the 3 cases: `( exp )` and
`var`.

adds implementation of `exp` for the case where you have a `prefixexp`