feat: fixes nested do .. end loops
fixes the nesting of do end loops, it appears i was actually adding
things wrong when i was trying to proces the inner statements of the do
.. end loop. i would try and smartly insert the new do .. end block into
the `block_stack` or the `working_block` (based on stack size) and then
right below it in the code take the `phrase` and add it to the
`block_stack` or the `working_block`. this ment that a rogue `end`
token was getting added and when we were trying to reduce everything
down to statements, it would error because `end` shouldn't exist at this
phase.