refactor: again, reworking it. done scanner
decided there was probably a better way to do this so decided to go for a
refactor based on everything i've learned thus far. the goals for this
refactor are (1) better error / code reference management (2) not so
complex parsing.

the rework of the scanner introduces (maybe) better code referencing,
instead of doing the `CodeWrap` method of before where we explode it
everytime we mutate the `Token`, we are now going to use a `CodeRef`
which is going to stay with the `Token` forever, which will reference
the raw_code start and stop as well as the line number.

cleaned up the error handling printing code to make it a little cleaner,
didn't change the content that is displayed.