feat: adds if-then-loops
adds the if-then loops into the mix, seems to work fine??

src/element.rs
- implements FromIterator for a CodeToken -> a CodeElement now, making it
easier to used drain and collect between the two types.
- adds `pretty_print`, which i needed because i was having an issue with
a more complex element structure (the if-then-loop) and wanted to see
exactly what was inside the element. put it inside the 'statement' check
as a feature for the `dev-testing` feature so i have it in the future
if something doesn't match to a statement pattern.

src/parser.rs
- adds the `process_if_loop` function so it now can process ifloops,

src/token.rs
i noticed that there are some issues waiting to rear their heads .. the
current way it gets the next token for these complexer structures (if,
do, while blocks) is that it checks for the ending token of a starting
token and then counts the nest, so it returns everything inside correctly.

when working on this one i noticed that i haven't been updating the
`matching_set` function and decided to add all of them, and things
stopped working. this was because some of the lua blocks share tokens
and make things confusing for doing it this way. so `while .. do .. end`
will trigger an error because it will see the `do` and assume that there
is a nested `do .. end` loop, and then not be able to find the 2nd `end`.

i need to fix this,