test: adds lua test suite for scanner
adds the lua test suite, but there are some issues. i commented out some
of the tests because they use characters that aren't part of the standard
US-EN character set, and cause issues with the `str::from_utf8` parsing.

need to figure out the right way to load files that might have other
encodings.

upon further investigation, the files that fail are `iso-8859-1` while the
files that pass are `us-ascii`.