fix: fixes the issue with nesting
changes the way it gets the `get_tokens_until_token` by changing the
`matching_token` function to return `Vec<Token>`. this way we need
to find the complete set of tokens for certain blocks, fixing the issues
where blocks using `do` would be looking for extra `end` and not find them.