fix: fixes issue with parser correctly parsing empty tables
the current table constructor code was really bad, it assumed that the
table construction ended at the end of the line so it screwed things up,
so if you did an inline table (or an empty table) it would get confused
and chop up the phrase all wrong and give you garbage.

now it actually looks for the ending moustache ('}') and uses that position
to correctly remove tokens.